Create components: testimonials, apply, footer and fix all
