+++ src/styles/components/apply.scss
@@ -1,0 +1,70 @@
+.apply {
+  font-family: $font-primary;
+  background-color: $white;
+  padding: 72px 0;
+
+  &__title {
+    font-weight: 600;
+    font-size: 44px;
+    line-height: 120%;
+    text-align: center;
+    margin-bottom: 24px;
+    color: $color-title;
+
+    @include on-tablet {
+      margin: 0 auto 25px;
+      display: block;
+      max-width: 450px;
+    }
+
+    @include on-big-tablet {
+      max-width: initial;
+    }
+  }
+
+  &__text {
+    font-family: $font-secondary;
+    font-weight: 600;
+    font-size: 22px;
+    line-height: 150%;
+    text-align: center;
+    margin-bottom: 56px;
+    color: $main-text-color;
+
+    @include on-big-tablet {
+      max-width: 700px;
+      margin: 0 auto 64px;
+    }
+  }
+
+  &__button {
+    font-weight: 600;
+    width: 100%;
+    font-size: 15px;
+    line-height: 22px;
+    text-align: center;
+    color: $white;
+  }
+
+  @include on-tablet {
+    &__inner {
+      padding: 200px 0;
+      background-image: url('../images/apply/apply-bg-small.png');
+      background-size: contain;
+      background-repeat: no-repeat;
+      background-position: center;
+    }
+
+    &__button {
+      display: block;
+      margin: 0 auto;
+      max-width: 264px;
+    }
+  }
+
+  @include on-big-tablet {
+    &__inner {
+      background-image: url('../images/apply/apply-bg-big.png');
+    }
+  }
+}
