+++ src/styles/components/aside-menu.scss
@@ -57,15 +57,7 @@
     position: relative;
 
     &::before {
-      content: '';
-      position: absolute;
-      bottom: -8px;
-      left: 50%;
-      width: 0;
-      height: 3px;
-      border-radius: 4px;
-      background-color: $color-link-bg;
-      transition: all 0.3s ease-in-out;
+      @include nav-link-before;
     }
 
     &:hover::before {
