+++ src/styles/components/button.scss
@@ -10,6 +10,7 @@ $amount: 15%;
   text-decoration: none;
   padding: 16px 32px;
   background: $color-link-bg;
+  display: inline-block;
   border-radius: 8px;
   color: inherit;
   transition: all 0.3s;
