+++ src/styles/components/header.scss
@@ -1,6 +1,6 @@
 .header {
   font-family: $font-primary;
-  background: $header-bg;
+  background: $dark-bg;
   padding: 10px 0 0;
   font-style: normal;
   line-height: 150%;
