+++ src/styles/components/services.scss
@@ -18,7 +18,7 @@
     }
   }
 
-  &__inner {
+  &__head {
     display: flex;
     flex-direction: column;
