+++ src/styles/components/testimonials.scss
@@ -1,0 +1,99 @@
+.testimonials {
+  font-family: $font-primary;
+  padding-bottom: 120px;
+  
+  &__subtitle {
+    font-weight: 700;
+    font-size: 13px;
+    line-height: 20px;
+    text-align: center;
+    letter-spacing: 3px;
+    text-transform: uppercase;
+    margin-bottom: 16px;
+    color: $subtitle-color;
+  }
+
+  &__title {
+    font-weight: 600;
+    font-size: 32px;
+    line-height: 150%;
+    margin-bottom: 48px;
+    color: $color-title;
+
+    @include on-tablet {
+      max-width: 190px;
+    }
+  }
+
+  &__inner {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+
+    @include on-big-tablet {
+      flex-direction: row;
+      justify-content: space-between;
+    }
+
+    @include on-desktop {
+      gap: 30px;
+    }
+  }
+
+  &__article {
+    background-color: $white;
+    border-radius: 16px;
+    text-align: center;
+    padding: 32px 37px;
+
+    img {
+      display: block;
+      max-width: 250px;
+      width: 100%;
+      height: auto;
+      margin: 0 auto 56px;
+      border-radius: 8px;
+    }
+  }
+
+  &__quote {
+    font-family: $font-secondary;
+    font-weight: 600;
+    font-size: 13px;
+    line-height: 150%;
+    text-align: center;
+    margin-bottom: 32px;
+    position: relative;
+    color: $main-text-color;
+
+    @include on-tablet {
+      margin: 0 auto 16px;
+      max-width: 350px;
+    }
+
+    &::before {
+      content: '';
+      display: inline-block;
+      position: absolute;
+      top: -35px;
+      left: 50%;
+      transform: translateX(-50%);
+      width: 21px;
+      height: 21px;
+      background-image: url('../images/testimonials/quote .svg');
+      background-size: auto;
+      background-repeat: no-repeat;
+      background-position: center;
+      margin-bottom: 10px;
+    }
+  }
+
+  &__author {
+    font-weight: 600;
+    font-size: 13px;
+    line-height: 150%;
+    text-align: center;
+    margin-bottom: 4px;
+    color: $color-title;
+  }
+}
