+++ src/styles/utils/_mixins.scss
@@ -74,3 +74,28 @@
     left: 0;
   }
 }
+
+@mixin nav-link-before {
+  content: '';
+  position: absolute;
+  bottom: -8px;
+  left: 50%;
+  width: 0;
+  height: 3px;
+  border-radius: 4px;
+  background-color: $color-link-bg;
+  transition: all 0.3s ease-in-out;
+}
+
+@mixin svg-mask($path) {
+  width: 32px;
+  height: 32px;
+  background-color: $white;
+  mask: url($path) no-repeat center;
+  mask-size: cover;
+  transition: background-color 0.3s ease-in-out;
+
+  &:hover {
+    background-color: $color-link-bg;
+  }
+}
