+++ src/styles/utils/_vars.scss
@@ -1,5 +1,5 @@
 $c-gray: #f2f6fa;
-$header-bg: #2c2c2c;
+$dark-bg: #2c2c2c;
 $white: #fff;
 $main-text-color: #6c788b;
 $color-text-dark: #334563;
